Fix CStr numeric constructors on 32-bit platforms. Remove unnecessary assignment operators.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@8926 3db68df2-c116-0410-a063-a993310a9797

+++ source/network/NetServer.cpp
@@ -710,7 +710,7 @@ CStrW CNetServerWorker::DeduplicatePlayerName(const CStrW& original)
 		if (unique)
 			return name;
 
-		name = original + L" (" + CStrW((unsigned long)id++) + L")";
+		name = original + L" (" + CStrW(id++) + L")";
 	}
 }
